+++ core/src/main/java/com/acme/calendar/core/KeycloakConstants.java
@@ -10,9 +10,9 @@ public class KeycloakConstants {
     public static final String ADMIN_USERNAME = "admin";
     public static final String ADMIN_PASSWORD = "admin@1234";
 
-    public static final String CLIENT_ID = "0c6da51a-08cc-45d3-8a44-f1c930aebad7";
-    public static final String CLIENT_NAME = "abhishek-khaiwale";
-    public static final String CLIENT_SECRET = "dSpdPWkChJ9MM3qTCCw9b2rz9OmTPt8O";
+    public static final String CLIENT_ID = "f0262919-c4d5-4f74-a106-827c39b7db34";
+    public static final String CLIENT_NAME = "calender-ui";
+    public static final String CLIENT_SECRET = "UCrcw4vQ671R4SXJGrklz8NqugxHSBpU";
 
     public static final String CREATE_USER = "/createUser";
     public static final String CREATE_GROUP = "/createGroup";
