--- services/rest/src/main/java/com/acme/calendar/service/views/CalendarView.java
@@ -1,6 +1,0 @@
-package com.acme.calendar.service.views;
-
-public class CalendarView {
-    public static class ExcludeEvents {}
-    public static class IncludeEvents extends ExcludeEvents {}
-}
